Work in progress

+++ src/tilemap.hpp
@@ -20,7 +20,7 @@ class tilemap final {
     std::shared_ptr<graphics::pixmap> _tileset;
     std::vector<geometry::rectangle> _tilesources;
     float_t _tilesize;
-    std::vector<std::vector<uint16_t>> _layers;
+    std::vector<std::vector<uint8_t>> _layers;
     std::shared_ptr<object> _target;
     geometry::rectangle _view;
 };
